% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\docType{class}
\name{epidecodeR-class}
\alias{epidecodeR-class}
\title{epidecodeR object - a S4 class object}
\description{
epidecodeR object - a S4 class object
}
\section{Slots}{

\describe{
\item{\code{t}}{data.frame.}

\item{\code{e}}{data.frame.}

\item{\code{eventcounts}}{numeric.}

\item{\code{grptables}}{list.}

\item{\code{grpcounts}}{integer.}

\item{\code{sign.test}}{data.frame.}
}}

