% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{get_eventcounts}
\alias{get_eventcounts}
\alias{get_eventcounts,epidecodeR-method}
\title{get_eventcounts method}
\usage{
get_eventcounts(object)

\S4method{get_eventcounts}{epidecodeR}(object)
}
\arguments{
\item{object}{epidecodeR object}
}
\value{
eventcounts
}
\description{
get_eventcounts method
}
\examples{
events<-system.file("extdata", "NOMO-1_ref_peaks.bed", package="epidecodeR")
deg<-system.file("extdata", "FTOi.txt", package="epidecodeR")
epiobj<-epidecodeR(events=events,deg=deg,pval=0.05,param=3,ints=c(2,4))
get_eventcounts(epiobj)
}
