% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{get_grpcounts}
\alias{get_grpcounts}
\alias{get_grpcounts,epidecodeR-method}
\title{get_grpcounts method}
\usage{
get_grpcounts(object)

\S4method{get_grpcounts}{epidecodeR}(object)
}
\arguments{
\item{object}{epidecodeR object}
}
\value{
grpcounts
}
\description{
get_grpcounts method
}
\examples{
events<-system.file("extdata", "NOMO-1_ref_peaks.bed", package="epidecodeR")
deg<-system.file("extdata", "FTOi.txt", package="epidecodeR")
epiobj<-epidecodeR(events=events,deg=deg,pval=0.05,param=3,ints=c(2,4))
get_grpcounts(epiobj)
}
