% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{get_grptables}
\alias{get_grptables}
\alias{get_grptables,epidecodeR-method}
\title{get_grptables method}
\usage{
get_grptables(object)

\S4method{get_grptables}{epidecodeR}(object)
}
\arguments{
\item{object}{epidecodeR object}
}
\value{
grptables
}
\description{
get_grptables method
}
\examples{
events<-system.file("extdata", "NOMO-1_ref_peaks.bed", package="epidecodeR")
deg<-system.file("extdata", "FTOi.txt", package="epidecodeR")
epiobj<-epidecodeR(events=events,deg=deg,pval=0.05,param=3,ints=c(2,4))
get_grptables(epiobj)
}
