% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{get_signtest}
\alias{get_signtest}
\alias{get_signtest,epidecodeR-method}
\title{get_signtest method}
\usage{
get_signtest(object)

\S4method{get_signtest}{epidecodeR}(object)
}
\arguments{
\item{object}{epidecodeR object}
}
\value{
signtest table
}
\description{
get_signtest method
}
\examples{
events<-system.file("extdata", "NOMO-1_ref_peaks.bed", package="epidecodeR")
deg<-system.file("extdata", "FTOi.txt", package="epidecodeR")
epiobj<-epidecodeR(events=events,deg=deg,pval=0.05,param=3,ints=c(2,4))
get_signtest(epiobj)
}
