% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{get_theoretical_table}
\alias{get_theoretical_table}
\alias{get_theoretical_table,epidecodeR-method}
\title{get_theoretical_table method}
\usage{
get_theoretical_table(object)

\S4method{get_theoretical_table}{epidecodeR}(object)
}
\arguments{
\item{object}{epidecodeR object}
}
\value{
theoretical_table
}
\description{
get_theoretical_table method
}
\examples{
events<-system.file("extdata", "NOMO-1_ref_peaks.bed", package="epidecodeR")
deg<-system.file("extdata", "FTOi.txt", package="epidecodeR")
epiobj<-epidecodeR(events=events,deg=deg,pval=0.05,param=3,ints=c(2,4))
get_theoretical_table(epiobj)
}
