% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{makeplot}
\alias{makeplot}
\title{Generate CDF plot using epidecodeR object generated using epidecodeR function}
\usage{
makeplot(obj, type, lim, title, xlab, ylab)
}
\arguments{
\item{obj}{epidecodeR object - epidecodeR object generated using epidecodeR function}

\item{type}{char - Type of CDF plot to generate; Accepted values 't': theoretical CDF plot; 'e': empirical CDF plot; 'both': Creates both theoretical and empirical plots.  Default: both}

\item{lim}{vector - Upper and lower limits of log2FC for X-axis}

\item{title}{char - Title of the plot}

\item{xlab}{char - X-axis label}

\item{ylab}{char - Y-axis label}
}
\value{
A CDF plot
}
\description{
Generate CDF plot using epidecodeR object generated using epidecodeR function
}
\examples{
p<-makeplot(epiobj, lim = c(-2,2), title = "H3K27ac", xlab = "siSDS3/WT")
p
OR
p<-makeplot(epiobj, type = 't', lim = c(-2,2),  title = "H3K27ac", xlab = "siSDS3/WT")
p
p<-makeplot(epiobj, lim=c(-7,7), type='e')
p
}
