% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{plot.test}
\alias{plot.test}
\title{Generates boxplot of distribution of log2FC of dysregulated genes and adjusted P value of signficance test of difference in mean log2FC between groups computed using one-way ANOVA test}
\usage{
\method{plot}{test}(obj, title, ylab)
}
\arguments{
\item{obj}{epidecodeR object - epidecodeR object generated using epidecodeR function}

\item{title}{char - Title of the plot}

\item{ylab}{char - Y-axis label}
}
\value{
Boxplot of distribution of log2FC of dysregulated genes between groups
}
\description{
Generates boxplot of distribution of log2FC of dysregulated genes and adjusted P value of signficance test of difference in mean log2FC between groups computed using one-way ANOVA test
}
\examples{
pt<-plot.test(epiobj, title = "H3K27ac", ylab = "siSDS3/WT")
pt
}
