% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{plot_test}
\alias{plot_test}
\title{Generates boxplot of distribution of log2FC of dysregulated genes and adjusted P value of signficance test of difference in mean log2FC between groups computed using one-way ANOVA test}
\usage{
plot_test(obj, title, ylab)
}
\arguments{
\item{obj}{epidecodeR object - epidecodeR object generated using epidecodeR function}

\item{title}{char - Title of the plot}

\item{ylab}{char - Y-axis label}
}
\value{
Boxplot of distribution of log2FC of dysregulated genes between groups
}
\description{
Generates boxplot of distribution of log2FC of dysregulated genes and adjusted P value of signficance test of difference in mean log2FC between groups computed using one-way ANOVA test
}
\examples{
events<-system.file("extdata", "con_peak.bed", package="epidecodeR")
deg<-system.file("extdata", "deg.txt", package="epidecodeR")
epiobj <- epidecodeR(events = events, deg = deg, pval=0.05, param = 3, ints=c(2,4))
plot_test(epiobj, title = "m6A mediated dysregulation after FTO inhibitor treatment", ylab = "log2FC")
}
