% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{plottingfunc}
\alias{plottingfunc}
\title{Title}
\usage{
plottingfunc(objdf, title, lim, xlab, ylab)
}
\arguments{
\item{objdf}{Data frame of the object from epidecodeR object}

\item{title}{Title of the plot}

\item{lim}{Limits of x-axis}

\item{xlab}{X-axis name}

\item{ylab}{Y-axis name}
}
\value{
returns a generic ggplot to be filled by makeplot function
}
\description{
Title
}
\examples{
\dontrun{
events<-system.file("extdata","NOMO-1_ref_peaks.bed", package="epidecodeR")
deg<-system.file("extdata", "FTOi.txt", package="epidecodeR")
epiobj<-epidecodeR(events=events,deg=deg,pval=0.05,param=3,ints=c(2,4))
plottingfunc(objdf=get_theoretical_table(epiobj), title="title", 
lim=c(-2,2), xlab="xlab", ylab="ylab")
}
}
\keyword{internal}
