% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidecodeR.R
\name{plottingfunc}
\alias{plottingfunc}
\title{Title}
\usage{
plottingfunc(objdf, title, lim, xlab, ylab)
}
\arguments{
\item{objdf}{Data frame of the object from epidecodeR object}

\item{title}{Title of the plot}

\item{lim}{Limits of x-axis}

\item{xlab}{X-axis name}

\item{ylab}{Y-axis name}
}
\value{
returns a generic ggplot to be filled by makeplot function
}
\description{
Title
}
\examples{
\dontrun{
plottingfunc(objdf=obj@t, title=title, lim=lim, xlab=xlab, ylab=ylab)
}
}
